refactor(core): move internal functionality to separated index
