+++ packages/core/src/internal.ts
@@ -1,0 +1,2 @@
+export * from './index';
+export * from './internal/process-declaration-functions';
